Fix paths and delete duplicate files, start to use separated amplitudelib

+++ src/fragmentation/fragmentation.hpp
@@ -7,7 +7,7 @@
  */
 
 #include <string>
-#include "../config.hpp"
+#include <tools/config.hpp>
 
 enum Hadron
 {
@@ -24,7 +24,7 @@ class FragmentationFunction
 {
     public:
         // D_{p->h}, x: long. mom. fraction, qs: scale (GeV)
-        virtual double Evaluate(Parton p, Hadron h, double x, double qs)=0;
+        virtual REAL Evaluate(Parton p, Hadron h, REAL x, REAL qs)=0;
         FragmentationFunction();
         virtual std::string GetString();
     private:
